bee_frontend/requirements.txt
streamlit
pymongo
streamlit_echarts
google-cloud
google-cloud-storage